+++ app/Http/Controllers/BarangPakaiController.php
@@ -17,6 +17,12 @@ class BarangPakaiController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct()
+    {
+        $this->middleware('dosen')->except(['index', 'show']);
+    }
+
     public function index()
     {
         //
@@ -30,12 +36,16 @@ class BarangPakaiController extends Controller
     public function create($lab)
     {
         $laboratorium = Laboratorium::find($lab);
-        $alat = Alat::all();
-        return view('v_barangpakai.create', [
-            'title' => 'Tambah Data Barang Pakai',
-            'laboratorium' => $laboratorium,
-            'alats' => $alat
-        ]);
+        if (auth()->user()->role == 'admin' || $laboratorium->user->id == auth()->user()->id) {
+            $alat = Alat::all();
+            return view('v_barangpakai.create', [
+                'title' => 'Tambah Data Barang Pakai',
+                'laboratorium' => $laboratorium,
+                'alats' => $alat
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -47,26 +57,30 @@ class BarangPakaiController extends Controller
     public function store(Request $request)
     {
         $laboratorium_id = $request->laboratorium_id;
-        $validatedData = $request->validate([
-            'kode' => 'required|unique:barang_pakais',
-            'nama' => 'required|max:255',
-            'laboratorium_id' => 'required',
-            'alat_id' => 'required',
-            'deskripsi' => 'required',
-            'keterangan' => 'required',
-            'upload' => 'required|image|mimes:jpg,jpeg,png|max:8000'
-        ]);
+        if (auth()->user()->role == 'admin' || $laboratorium_id == auth()->user()->id) {
+            $validatedData = $request->validate([
+                'kode' => 'required|unique:barang_pakais',
+                'nama' => 'required|max:255',
+                'laboratorium_id' => 'required',
+                'alat_id' => 'required',
+                'deskripsi' => 'required',
+                'keterangan' => 'required',
+                'upload' => 'required|image|mimes:jpg,jpeg,png|max:8000'
+            ]);
 
-        if ($request->file('upload')) {
-            $validatedData['upload'] = $request->file('upload')->store('barangpakai-images');
+            if ($request->file('upload')) {
+                $validatedData['upload'] = $request->file('upload')->store('barangpakai-images');
+            }
+
+            $validatedData['foto'] = $validatedData['upload'];
+            unset($validatedData['upload']);
+
+            BarangPakai::create($validatedData);
+
+            return redirect('/laboratorium/' . $laboratorium_id)->with('success', 'Tambah Data Barang Pakai Berhasil');
+        } else {
+            abort(403);
         }
-
-        $validatedData['foto'] = $validatedData['upload'];
-        unset($validatedData['upload']);
-
-        BarangPakai::create($validatedData);
-
-        return redirect('/laboratorium/' . $laboratorium_id)->with('success', 'Tambah Data Barang Pakai Berhasil');
     }
 
     /**
@@ -97,12 +111,16 @@ class BarangPakaiController extends Controller
     public function edit($id)
     {
         $barangpakai = BarangPakai::find($id);
-        $alat = Alat::all()->except([$barangpakai->alat_id]);
-        return view('v_barangpakai.edit', [
-            'title' => $barangpakai->nama,
-            'barangpakai' => $barangpakai,
-            'alats' => $alat
-        ]);
+        if (auth()->user()->role == 'admin' || $barangpakai->laboratorium->user->id == auth()->user()->id) {
+            $alat = Alat::all()->except([$barangpakai->alat_id]);
+            return view('v_barangpakai.edit', [
+                'title' => $barangpakai->nama,
+                'barangpakai' => $barangpakai,
+                'alats' => $alat
+            ]);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -115,32 +133,36 @@ class BarangPakaiController extends Controller
     public function update(Request $request, $id)
     {
         $barangpakai = BarangPakai::find($id);
-        $rules = [
-            'nama' => 'required|max:255',
-            'alat_id' => 'required',
-            'deskripsi' => 'required',
-            'keterangan' => 'required',
-            'upload' => 'required|image|mimes:jpg,jpeg,png|max:8000'
-        ];
+        if (auth()->user()->role == 'admin' || $barangpakai->laboratorium->user->id == auth()->user()->id) {
+            $rules = [
+                'nama' => 'required|max:255',
+                'alat_id' => 'required',
+                'deskripsi' => 'required',
+                'keterangan' => 'required',
+                'upload' => 'required|image|mimes:jpg,jpeg,png|max:8000'
+            ];
 
-        if ($request->kode != $barangpakai->kode) {
-            $rules['kode'] = 'required|unique:barang_pakais';
-        }
-
-        $validatedData = $request->validate($rules);
-
-        if ($request->file('upload')) {
-            if ($request->oldImage) {
-                Storage::delete($request->oldImage);
+            if ($request->kode != $barangpakai->kode) {
+                $rules['kode'] = 'required|unique:barang_pakais';
             }
-            $validatedData['upload'] = $request->file('upload')->store('barangpakai-images');
+
+            $validatedData = $request->validate($rules);
+
+            if ($request->file('upload')) {
+                if ($request->oldImage) {
+                    Storage::delete($request->oldImage);
+                }
+                $validatedData['upload'] = $request->file('upload')->store('barangpakai-images');
+            }
+
+            $validatedData['foto'] = $validatedData['upload'];
+            unset($validatedData['upload']);
+
+            BarangPakai::where('id', $barangpakai->id)->update($validatedData);
+            return redirect('/laboratorium/' . $barangpakai->laboratorium_id)->with('success', 'Ubah Data Barang Pakai Berhasil');
+        } else {
+            abort(403);
         }
-
-        $validatedData['foto'] = $validatedData['upload'];
-        unset($validatedData['upload']);
-
-        BarangPakai::where('id', $barangpakai->id)->update($validatedData);
-        return redirect('/laboratorium/' . $barangpakai->laboratorium_id)->with('success', 'Ubah Data Barang Pakai Berhasil');
     }
 
     /**
@@ -152,14 +174,18 @@ class BarangPakaiController extends Controller
     public function destroy($id)
     {
         $barangpakai = BarangPakai::find($id);
-        $laboratorium = Laboratorium::find($barangpakai->laboratorium_id);
+        if (auth()->user()->role == 'admin' || $barangpakai->laboratorium->user->id == auth()->user()->id) {
+            $laboratorium = Laboratorium::find($barangpakai->laboratorium_id);
 
-        if ($barangpakai->foto) {
-            Storage::delete($barangpakai->foto);
+            if ($barangpakai->foto) {
+                Storage::delete($barangpakai->foto);
+            }
+
+            barangpakai::destroy($barangpakai->id);
+
+            return redirect('/laboratorium/' . $laboratorium->id)->with('success', 'Data Barang Pakai Berhasil Dihapus');
+        } else {
+            abort(403);
         }
-
-        barangpakai::destroy($barangpakai->id);
-
-        return redirect('/laboratorium/' . $laboratorium->id)->with('success', 'Data Barang Pakai Berhasil Dihapus');
     }
 }
